+++ extensions/jui/CHANGELOG.md
@@ -4,7 +4,10 @@ Yii Framework 2 jui extension Change Log
 2.0.0-rc under development
 --------------------------
 
-- no changes in this release.
+- Chg #1551: Jui datepicker has a new property `$dateFormat` which is used to set the clientOption `dateFormat`.
+   The new property does not use the datepicker formatting synax anymore but uses the same as the `yii\i18n\Formatter`
+   class which is the ICU syntax for date formatting, you have to adjust all your DatePicker widgets to use
+   the new property instead of setting the dateFormat in the clientOptions (cebe)
 
 
 2.0.0-beta April 13, 2014